Filetype: Rewrite C compile functions using term.
There is a weird bug with term and running C files which requires
hitting enter once after calling a function. This means you cannot
simply call a function. You must make the function a command then call
the command, and scripting the enter press twice. This weirdness is why
this function is so different from the "CompileJava" function.