Plugin: Add more robust include characters for rainbow parens.
Originally I wanted to be able to highlight '/=' but figured out this
overrides the syntax for a `//` comment.

The solution is to make the include characters much more specific. If you want
to highlight the C assignment operator `=` then add ` = ` to the include array.
Or in my case if you want to highlight `/=` explicitly add ` /= ` to the
include array. You must be sure to use single quotes and escape `\`, `.`, and
`*` characters.