Plugin: Run rainbow parens for most filetypes.
Ideally RainbowParentheses should run is every filetype but a select few but
due to limitations with the FileType autocommand it is not possible to exclude
filetypes.

The solution is to run a function in every file that checks an array that
contains filetypes where RainbowParentheses should be turned off and proceed
accordingly.